Refactor data generation script to enhance sample size and implement graceful exit handling
